refactor(prefabs): reorg prefabs into their own folders
