+++ src/prefabs/button/button.scss
@@ -1,4 +1,4 @@
-@import './variables';
+@import '../../variables';
 
 // Mixin (used by the spark pill and the button itself) to define base
 // background and border colors/styles.
