+++ src/prefabs/button/button.tsx
@@ -1,9 +1,9 @@
 import * as Mixer from '@mcph/miix-std';
 import { Component, h } from 'preact';
 
-import { gamepad } from './alchemy/Gamepad';
-import { PreactControl } from './alchemy/preact/index';
-import { classes, css } from './alchemy/Style';
+import { gamepad } from '../../alchemy/Gamepad';
+import { PreactControl } from '../../alchemy/preact/index';
+import { classes, css } from '../../alchemy/Style';
 
 /**
  * SparkPill is the component that shows the spark cost above a button.
