+++ src/prefabs/default-scene/scene.tsx
@@ -1,10 +1,12 @@
 import * as Mixer from '@mcph/miix-std';
 import { h } from 'preact';
 
-import { PreactScene } from './alchemy/preact';
+import { PreactScene } from '../../alchemy/preact';
 
 /**
- * Default "stub" scene.
+ * This is the basic Interactie scene. Your scene will contain many controls,
+ * inside the <Layout>. You can add custom scene-wide styling and behaviors,
+ * like background images, here.
  */
 @Mixer.Scene({ default: true })
 export class DefaultScene extends PreactScene<{}> {
