+++ src/prefabs/joystick/joystick.scss
@@ -1,3 +1,5 @@
+@import '../../variables';
+
 .mixer-joystick {
     position: absolute;
     top: 0;
